+++ examples/01-scalar/src/bindings.zig
@@ -3,5 +3,5 @@ const scalar = @import("scalar");
 
 pub const bindings = zigo.define(.{
     .root = scalar,
-    .functions = .{.{ .path = "root.add" }},
+    .functions = .{.{ .path = "root.add", .covers = "root.subtract" }},
 });